chore: rename lib/common/stdvector.h => lib/common/StdVector.h
