+++ src/lib/base/PriorityQueue.h
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "common/stdvector.h"
+#include "common/StdVector.h"
 
 #include <algorithm>
 #include <functional>
