+++ src/lib/base/String.cpp
@@ -7,7 +7,7 @@
  */
 
 #include "base/String.h"
-#include "common/stdvector.h"
+#include "common/StdVector.h"
 
 #include <algorithm>
 #include <cctype>
