+++ src/lib/deskflow/KeyMap.h
@@ -10,7 +10,7 @@
 #include "base/String.h"
 #include "common/StdMap.h"
 #include "common/StdSet.h"
-#include "common/stdvector.h"
+#include "common/StdVector.h"
 #include "deskflow/KeyTypes.h"
 
 #ifdef TEST_ENV
