+++ src/lib/deskflow/OptionTypes.h
@@ -8,7 +8,7 @@
 #pragma once
 
 #include "base/EventTypes.h"
-#include "common/stdvector.h"
+#include "common/StdVector.h"
 
 //! Option ID
 /*!
