+++ src/lib/deskflow/ProtocolUtil.cpp
@@ -7,7 +7,7 @@
 
 #include "deskflow/ProtocolUtil.h"
 #include "base/Log.h"
-#include "common/stdvector.h"
+#include "common/StdVector.h"
 #include "deskflow/ProtocolTypes.h"
 #include "deskflow/XDeskflow.h"
 #include "io/IStream.h"
