+++ src/lib/io/StreamBuffer.h
@@ -9,7 +9,7 @@
 
 #include "base/EventTypes.h"
 #include "common/StdList.h"
-#include "common/stdvector.h"
+#include "common/StdVector.h"
 
 //! FIFO of bytes
 /*!
