+++ src/lib/platform/MSWindowsClipboard.h
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "common/stdvector.h"
+#include "common/StdVector.h"
 #include "deskflow/IClipboard.h"
 #include "platform/MSWindowsClipboardFacade.h"
