+++ src/lib/platform/MSWindowsKeyState.h
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "common/stdvector.h"
+#include "common/StdVector.h"
 #include "deskflow/KeyState.h"
 
 #include <string>
