+++ src/lib/platform/OSXScreen.h
@@ -9,7 +9,7 @@
 
 #include "base/EventTypes.h"
 #include "common/StdMap.h"
-#include "common/stdvector.h"
+#include "common/StdVector.h"
 #include "deskflow/DragInformation.h"
 #include "deskflow/PlatformScreen.h"
 #include "platform/OSXClipboard.h"
