+++ src/lib/platform/XWindowsClipboard.h
@@ -9,7 +9,7 @@
 
 #include "common/StdList.h"
 #include "common/StdMap.h"
-#include "common/stdvector.h"
+#include "common/StdVector.h"
 #include "deskflow/ClipboardTypes.h"
 #include "deskflow/IClipboard.h"
