+++ src/lib/platform/XWindowsScreen.h
@@ -9,7 +9,7 @@
 
 #include "arch/Arch.h"
 #include "common/StdSet.h"
-#include "common/stdvector.h"
+#include "common/StdVector.h"
 #include "deskflow/ClientArgs.h"
 #include "deskflow/KeyMap.h"
 #include "deskflow/PlatformScreen.h"
