+++ src/lib/platform/XWindowsUtil.h
@@ -9,7 +9,7 @@
 
 #include "base/EventTypes.h"
 #include "common/StdMap.h"
-#include "common/stdvector.h"
+#include "common/StdVector.h"
 
 #include <string>
