+++ src/lib/server/ClientProxy1_5.h
@@ -7,7 +7,7 @@
 #pragma once
 
 #include "base/Stopwatch.h"
-#include "common/stdvector.h"
+#include "common/StdVector.h"
 #include "server/ClientProxy1_4.h"
 
 class Server;
