+++ src/lib/server/Server.h
@@ -12,7 +12,7 @@
 #include "base/Stopwatch.h"
 #include "common/StdMap.h"
 #include "common/StdSet.h"
-#include "common/stdvector.h"
+#include "common/StdVector.h"
 #include "deskflow/Clipboard.h"
 #include "deskflow/ClipboardTypes.h"
 #include "deskflow/DragInformation.h"
